ddd, Oct 10/25/2022, YYYY, 10:00 A +0

+++ pset9/finance/app.py
@@ -209,7 +209,7 @@ def sell():
             quote = lookup(symbol)
             id = session["user_id"]
             cash = db.execute("SELECT cash FROM users WHERE id = ?", id)
-            cost = int(shares) * float(quote["price"])
+            cost = float(shares) * float(quote["price"])
             db.execute("INSERT INTO transactions (user_id, type, symbol, quantity, price, cost) VALUES (?, ?, ?, ?, ?, ?)", id, 'SELL', symbol, shares, quote["price"], cost)
             remcash = float(cash[0]["cash"]) + cost
             db.execute("UPDATE users SET cash = ? WHERE id = ?", 10000, id)
